Fix conflicts after merge with develop branch
