+++ ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/activity-blocks/modalActivityBlock/modalActivityBlock.component.scss
@@ -69,7 +69,7 @@
 
             & > button {
                 height: var(--modal-activity-block-delete-dialog-button-height, 40px);
-                width: var(--modal-activity-block-delete-dialog-button-width, 140px);
+                width: var(--modal-activity-block-delete-dialog-button-width, 9rem);
             }
         }
     }
@@ -121,16 +121,18 @@
             padding: 27px 45px 20px 45px;
         }
 
-        &:after { // arrow
-            content: " ";
-            position: relative;
-            height: 0;
-            right: 50%;
-            bottom: -100%;
-            border-top: 10px solid white;
-            border-right: 10px solid transparent;
-            border-left: 10px solid transparent;
-            border-bottom: none;
+        @media (min-width: 768px) {
+            &:after { // arrow
+                content: " ";
+                position: relative;
+                height: 0;
+                right: 50%;
+                bottom: -100%;
+                border-top: 10px solid white;
+                border-right: 10px solid transparent;
+                border-left: 10px solid transparent;
+                border-bottom: none;
+            }
         }
     }
 
@@ -140,5 +142,9 @@
 
     .ctrl-btn:focus {
         outline: -webkit-focus-ring-color auto 1px;
+    }
+
+    mat-spinner {
+        margin: 10% 40%;
     }
 }
