--- ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/modalActivityBlock/modalActivityBlock.component.html
@@ -1,82 +1,0 @@
-<mat-card *ngFor="let instance of block.instances" class="modal-activity-block__card">
-  <mat-card-content class="modal-activity-block__card-content">
-    <div>
-      <mat-card-title class="modal-activity-block__card-title">
-        {{instance.activityName}}
-      </mat-card-title>
-      <mat-card-subtitle *ngIf="instance.activityDescription"
-                         class="modal-activity-block__card-subtitle">
-        {{instance.activityDescription}}
-      </mat-card-subtitle>
-
-      <div *ngIf="instance.statusCode !== 'CREATED'"
-           class="modal-activity-block__success-message">
-        <mat-icon inline>check_circle</mat-icon>
-        <span [innerHTML]="'SDK.ModalActivityBlock.NumQuestionsAnswered'
-          | translate: { number: instance.numQuestionsAnswered }"></span>
-      </div>
-    </div>
-
-    <mat-card-actions class="modal-activity-block__card-buttons">
-      <button mat-mini-fab
-              *ngIf="!instance.readonly"
-              class="button_primary"
-              (click)="openEditDialog(instance)">
-        <img src="assets/images/pencil.svg"
-             [alt]="'SDK.EditButton' | translate">
-      </button>
-      <button mat-mini-fab
-              *ngIf="instance.canDelete"
-              #delete_button
-              class="button_secondary"
-              (click)="openDeleteDialog(instance)">
-        <img src="assets/images/trash.svg"
-             [alt]="'SDK.DeleteButton' | translate">
-      </button>
-    </mat-card-actions>
-  </mat-card-content>
-</mat-card>
-
-<ng-template #edit_dialog>
-  <div mat-dialog-title class="modal-activity-block__edit-dialog-title">
-    <div>{{currentActivityInstance?.activityName}}</div>
-    <button mat-icon-button mat-dialog-close class="modal-activity-block__close-button">
-      <mat-icon>close</mat-icon>
-    </button>
-  </div>
-  <mat-dialog-content class="modal-activity-block__edit-dialog-content">
-    <mat-spinner *ngIf="!activityForm; else section"></mat-spinner>
-    <ng-template #section>
-      <ddp-activity-section [section]="activityForm?.sections[0]"
-                            [readonly]="readonly"
-                            [studyGuid]="studyGuid"
-                            [activityGuid]="currentActivityInstance.instanceGuid"
-                            [validationRequested]="validationRequested"
-      ></ddp-activity-section>
-    </ng-template>
-
-    <button mat-flat-button mat-dialog-close
-            class="button_primary modal-activity-block__save_activity"
-            [innerHTML]="'SDK.SaveButton' | translate">
-    </button>
-  </mat-dialog-content>
-</ng-template>
-
-<ng-template #delete_dialog>
-  <div class="modal-activity-block__delete-dialog">
-    <span class="modal-activity-block__delete-dialog-title"
-          [innerHTML]="'SDK.ModalActivityBlock.ConfirmDeletion' | translate">
-    </span>
-    <div class="modal-activity-block__delete-dialog-buttons">
-      <button mat-flat-button
-              color="warn"
-              (click)="deleteActivityInstance()"
-              [innerHTML]="'SDK.DeleteButton' | translate">
-      </button>
-      <button mat-stroked-button
-              mat-dialog-close
-              [innerHTML]="'SDK.CancelButton' | translate">
-      </button>
-    </div>
-  </div>
-</ng-template>
